+++ api/createPipe.ts
@@ -2,14 +2,17 @@ import { Signal, Sink, Source, Talkback } from "strict-callbag"
 import { subscribe, Subscription } from "./subscribe"
 
 interface Callbacks<A, EI, EO> {
-  onStart: (sub: Subscription) => void
-  onData: (sub: Subscription, data: A) => void
-  onEnd: (err?: EI) => void
+  onStart: (this: void, sub: Subscription) => void
+  onData: (this: void, sub: Subscription, data: A) => void
+  onEnd: (this: void, err?: EI) => void
 
-  onRequest: (sub: Subscription) => void
-  onAbort: (err?: EO) => void
+  onRequest: (this: void, sub: Subscription) => void
+  onAbort: (this: void, err?: EO) => void
 
-  talkbackOverride?: (original: Talkback<any>) => Talkback<any>
+  talkbackOverride?: (
+    this: void,
+    original: Talkback<unknown>,
+  ) => Talkback<unknown>
 }
 
 /**
@@ -17,7 +20,7 @@ interface Callbacks<A, EI, EO> {
  */
 export const createPipe = <A, EI, EO = never>(
   source: Source<A, EI>,
-  sink: Sink<any, EI, EO>,
+  sink: Sink<any, EI, EO>, // eslint-disable-line
   {
     onStart,
     onData,