+++ api/interval.ts
@@ -1,0 +1,5 @@
+import cbInterval from "callbag-interval"
+import { Source } from "strict-callbag"
+
+export const interval = (ms: number): Source<number, never> =>
+  cbInterval(ms) as Source<number, never>