+++ api/toAsyncIterable.ts
@@ -1,5 +1,6 @@
 import cbTAI from "callbag-to-async-iterable"
 import { Source } from "strict-callbag"
 
-export const toAsyncIterable = <A>(self: Source<A, any>): AsyncIterable<A> =>
-  cbTAI(self)
+export const toAsyncIterable = <A>(
+  self: Source<A, unknown>,
+): AsyncIterable<A> => cbTAI(self)