+++ modules.d.ts
@@ -11,7 +11,7 @@ declare module "callbag-start-with" {
 declare module "callbag-to-async-iterable" {
   import { Source } from "strict-callbag"
 
-  const toAsyncIterable: <A>(self: Source<A, any>) => AsyncIterable<A>
+  const toAsyncIterable: <A>(self: Source<A, unknown>) => AsyncIterable<A>
 
   export default toAsyncIterable
 }
@@ -23,4 +23,24 @@ declare module "callbag-from-obs" {
   const fromObservable: <A>(obs: InteropObservable<A>) => Source<A, unknown>
 
   export default fromObservable
+}
+
+declare module "callbag-buffer" {
+  import { Source } from "strict-callbag"
+
+  const buffer: (
+    seperator: Source<unknown, never>,
+  ) => <A, E>(source: Source<A, E>) => Source<A, E>
+
+  export default buffer
+}
+
+declare module "callbag-buffer-time" {
+  import { Source } from "strict-callbag"
+
+  const bufferTime: (
+    duration: number,
+  ) => <A, E>(source: Source<A, E>) => Source<A, E>
+
+  export default bufferTime
 }